Frontend Improvements
The pages now all use a shared style sheet. I plan to keep this the same in the final version by getting the backend to insert the data from the html before it is sent to the client or by even getting a runner to create the new versions of the files before they are sent through github or something similar. I have updated the Readme with some ideas and an upadted roadmap.