Articles work now!
In this update lots has changed. The reading side of the entire article system is done. Users can be presented with articles to read and click on them and taken to the full article. All the formatting is correct as well. Next will probably be adding articles then finally I will add a search box