Interactions complete and more...
Likes now work on all articles. I can now write articles to the database and then they show up on the site. There is also a page to show articles as a nav page. Automatic deployments through an endpoint are in the works. There is a problem with the mysql database on the planned host that is getting in the way of any testing with that. Comments are off to a start with the barebones for the front end complete.