Data Validation
This update while including some little extras like changes to the nav. Mostly focuses on data validation. Now if you don't provide an email in your comment it doesn't crash the webserver. I have also stopped script tags from being able to be put into a comment. This should then be used on the frontend where a message could be displayed. This will probably come in a fix to comments to stop the same comment being posted multiple times