QoL improvement + comment likes
This update comes with changes to the management interfaces. This means there is a login screen for the management pages. There is also images now for articles, as well as there being likes on comments. Plus the last comment that you posted will show up with "(You)" at the end of the name.