User Auth
There is now user authentication. This is soon going to mean that you will have to create an account to like articles and comments plus you will need one to write comments. For the early adopters who create accounts will see their username in the top left of the page