Removed non-needed queries
Certain routes such as the comments got data from the database such as the comments and the id of the user who wrote them. Then it used to loop through an get all the usernames. I have now made one query that removes the need for this loop and the extra queries.