User moderation
There is now a page that shows all the accounts and allows the sudo accounts to remove inappropriate profile pictures. If a user continues to change their profile picture back then the admin can disable that account from changing their profile picture. The user is still able to remove their profile picture. An admin is also able to disable a user from being able to comment. Comments can also be deleted by the user that posted them and by admins.