Locking accounts
This update allows admins to press a button and for the account specified to have any sessions that are current to be logged out. Then if a user that has their account locked tries to log back in they get a message that tells them that the account is locked. This update also shows when an account was last seen